+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Include fixed due date schedule list in loan-policy form. Fixes UICIRC-27.
 * Show empty schedules in new Fixed Due Date Schedule form. Fixes UICIRC-30.
 * Add empty schedule rows to top of list. Fixes UICIRC-29.
+* Fix loan policy form. Fixes UICIRC-34.
 
 ## [1.1.1](https://github.com/folio-org/ui-circulation/tree/v1.1.1) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.1.0...v1.1.1)